barycentric coordinates implementation
