+++ with_gpu.out
@@ -1,13 +1,13 @@
 ## SLURM PROLOG ###############################################################
-##    Job ID : 12387519
+##    Job ID : 12606737
 ##  Job Name : cuda.sh
-##  Nodelist : gpu1402
+##  Nodelist : gpu2007
 ##      CPUs : 1
 ##  Mem/Node : 20480 MB
 ## Directory : /oscar/home/nxu4/apma2822/apma2822B_final_project_vignesh_neil
-##   Job Started : Wed Dec 13 22:02:46 EST 2023
+##   Job Started : Fri Dec 22 18:05:27 EST 2023
 ###############################################################################
-Wed Dec 13 22:02:46 2023       
+Fri Dec 22 18:05:27 2023       
 +---------------------------------------------------------------------------------------+
 | NVIDIA-SMI 535.54.03              Driver Version: 535.54.03    CUDA Version: 12.2     |
 |-----------------------------------------+----------------------+----------------------+
@@ -15,8 +15,8 @@ Wed Dec 13 22:02:46 2023
 | Fan  Temp   Perf          Pwr:Usage/Cap |         Memory-Usage | GPU-Util  Compute M. |
 |                                         |                      |               MIG M. |
 |=========================================+======================+======================|
-|   0  NVIDIA TITAN RTX               Off | 00000000:61:00.0 Off |                  N/A |
-| 41%   28C    P8              12W / 280W |     19MiB / 24576MiB |      0%      Default |
+|   0  Quadro RTX 6000                Off | 00000000:40:00.0 Off |                  Off |
+| 33%   31C    P8              24W / 260W |     22MiB / 24576MiB |      0%      Default |
 |                                         |                      |                  N/A |
 +-----------------------------------------+----------------------+----------------------+
                                                                                          
@@ -25,6 +25,6 @@ Wed Dec 13 22:02:46 2023
 |  GPU   GI   CI        PID   Type   Process name                            GPU Memory |
 |        ID   ID                                                             Usage      |
 |=======================================================================================|
-|    0   N/A  N/A     22179      G   X                                            17MiB |
+|    0   N/A  N/A     22430      G   X                                            20MiB |
 +---------------------------------------------------------------------------------------+
 done
